fix warnings

+++ src/stellar_api.h
@@ -110,6 +110,6 @@ void print_int(int64_t l, char *out);
 void print_uint(uint64_t l, char *out);
 
 /** base64 encoding function used to display managed data values */
-void base64_encode(const uint8_t *data, size_t inLen, char *out);
+void base64_encode(const uint8_t *data, int inLen, char *out);
 
 #endif
